Refine the styles for some of the dock widgets
This is unfortunately necessary due to an oversight of mine from the
beginning: it may have been easier to incorporate all of the buttons
into the same grid, rather than having a top-level body-grid, and a
numberpad "subgrid", which doesn't necessarily agree with the
body-grid.

However, in my opinion, the final visual effect is still decent. :)