Implement decimal-point functionality
The decimal point sets a Boolean that tells the number event-listeners
to increase the scale with each click, giving the illusion that digits
are being "appended" to the end of the currently displayed figure.

What's nice about the current implementation is that it avoids
completely the problem of multiple decimal points that the project
description mentions; a number like 12.3.56.5 is impossible to enter
simply by virtue of how the code is written.