Refactor decimal-point-active flag into factory-object
The factory object 'calcState' now manages state for the calculator
application. This is mainly to keep global-variable use under control.

The DOM references have also all been moved to the top of the
file. The main reason for this is to let them appear within the scope
of 'initCalcState'. This way, the function can freely use whatever DOM
references it should need going forward.