Introduce a special value representing zero
Often, it's convenient to think of zero as, for example, a common
value to reset to. This commit introduces 'zeroValue' in preparation
for when the user presses certain buttons that require the display to
be "zeroed" when a new value needs to be edited for introduction into
the current calculation.