Implement arithmetic-button functionality
This commit introduces basic arithmetic-calculation functionality.

The basic idea would've been to introduce, for the digit buttons, a
one-shot event handler that should fire _before_ the usual one that
types the digits in. However, prepending event handlers isn't
possible, so the trick here is to use event capturing on the parent
".area-numberpad" container (whose DOM reference we're now forced to
isolate), placing our one-shot event handler on this element instead.

Also, previous attempts of this code mistakenly called 'writeValue' on
a raw number, without first converting it to a value object. Hence, it
was necessary to generalize 'readValue' to accept any string
parameter, which was done (in a previous commit) precisely to
prepare for this commit.