Implement clear-entry functionality via CE button
I've added a new method to the editor which clears a buffer. Based on
how this method works (it only returns the empty string), for now the
buffer need not be passed into it as a parameter.

There may be a case in the future where a cleared display has some
default text such as "0.".