Assign digits and decimal point to their respective keys
The approach taken here, unfortunately, isn't quite seamless - I need
to artifically trigger a click on the 'numberPad' container, so that
the one-shot event listener added by 'flushArithmetic' is still fired,
as it normally would be when the user is interacting with the
calculator via the mouse. This isn't quite as easy to do purely off of
the keyboard, since my keyboard events are captured only by the
'document' element.