Eliminate binary operators, dismantle 'editor' into components
1. The decimalPointState object, via its member closures, manages the
state of the decimal point (to avoid the need to represent this state
as a global variable). It serves no other purpose.

2. The original 'editor' object functionality has been broken down
into a set of functions: 'insertIntoDisplay', backspaceDisplay', and
'clearDisplay'. Since these don't manage any state, they don't belong
in an object.

3. All binary operation functionality (addition, etc.) has been
erased, so that I can redesign the buffering and flushing of tokens
taken from the display.

4. Event listeners that don't rely on this buffer, or otherwise any
persistent state (for example, the ones controlling the unary
operators and memory buttons) have been kept.