Make default cleared-display show '0' instead of being blank
The motivation for this decision is based on the fact that toggling
the sign on a blank display generates NaN. Now, this can't happen.

Of course, a little code needs to be added handling the case for when
user-input must overwrite this initial '0'.

It's also more consistent with how the example project, linked to in
the project description, handles this.