Handle divide-by-zero and negative-square-root exceptions
I've chosen the snarky message (see problem description) to be the
word "no". To keep things simple (e.g. avoid using promises/async),
the 'calculate' function is potentially ready to accept a first
argument of "no", which is given the value of 0.

This should be the only place where "no" gets introduced as a value,
and so it need not be handled anywhere else.